Reduce UI latency by doing less frequent lookups
The frequency of mouseover events was causing the updates to lag behind
the mouse pointer by a noticeable amount. The most expensive part of the
mouseover event is looking up the data point.

This change throttles the updates with a 50ms window. This is enough to
still feel instant but significantly reduce the overhead.