Consolidate javascript into a single output file
The initial goal with the separate `render_report.js` output file was to
have a lazily loaded JS resource for the report renderer which could be
eagerly cached by the browser.

It wasn't proving easy to load this in an async manner, because of the
requirement for jQuery, and the way that interacted with the load order
of scripts.