Implement stacked area graph
Some minor tweaks to the graph logic allows stacked area graphs to be
built. The primary difference with a stacked graph is that the series
order becomes important, and the descriptive text does not necessarily
match the metric which is drawn.

While d3 supports "proper" stacked graphs, there's no discernible
difference and this approach was easier to implement.