Use entity_ids instead of references to IdP / SP
While the foreign key was great for database normalisation, it causes a
problem where the entity ID isn't yet known by the reporting service
when an event is received.

There's a related issue where, if an IdP or SP is deleted from FR, the
resulting delete in reporting service will never succeed because the
foreign key prevents it (and we don't want to cascade the delete,
because a bug could cause a lot of lost event records).