Extend request timeout to 5 minutes
Report generation for large datasets is timing out in users browsers
resulting in a 502 response.

This initial change will allow the ruby process more time to generate
the report response. A change to the associated Apache proxy connection
will also be added in AAF Ansible.

This is a partial solution for #148.