Optimize utilization report query
The utilization report was loading every session object in the time
range, which was pretty dramatically incorrect behaviour given the size
of the table in production. In a 1-year time period, over 1 million rows
are generated.

This change avoids loading any of the objects directly, and instead
performs a join query with a `count(*)`.