Refactor indexes to optimize utilization reports
The new query format did a lot to reduce allocations in the Ruby
process, but the query plan generated by MySQL was not optimal due to
the structure of our indexes. There are two parts to this change:

- The unique index on (phase, unique_id) was being incorrectly targeted
  by the query optimizer, ignoring the index on `timestamp` entirely.
  Switching the order of columns in that index retains the unique
  constraint while avoiding that index interfering in query optimization
  (we *never* want to query on unique_id, except to enforce uniqueness).
- The index on `timestamp` was only covering part of the query, and a
  decent speed boost could be attained by adding `phase` into that
  index. Given that `phase` is almost always part of the query, it can
  be first in the index.

Testing before/after these changes, against a local database, shows that
the query time drops from 3 minutes to approximately half a second.