SessionSource: hide Subscribe form if params reqd but not present
Conditionally hide the Subscribe form on reports that need @source.

Otherwise, following the Subscribe links directly from the landing page
would lead to validation errors.

Such hiding is already done for reports that need a target - the Subscribe form
is only only shown when the report is rendered.

This is done by wrapping the Subscribe form rendering code with `- if @data`

This is inline with the instructions displayed on the Subscriptions page:

  "If you couldn't find the Subscribe button on report page you may need to generate a report first"