SessionSource: refine defaults in set_source
On FederationReportsController, for public reports, set_source sets @source to
a default value even when no parameter is provided.

This is needed for the DailyDemand and FederatedSessionsReport public reports,
which do not have a form where the source would be set.

However, calling set_source as a before_action for all actions on
FederationReportsController caused @source to be set also when generating
a FederationGrowthReport, where source would then be included in the
Subscribe links.

Add an 'only' clause on the set_source before_action to prevent this from happening.

And add a comment in set_source to explain the need for a default value.