SessionSource: spec: AR factory: chg dflt report_class
Change the default report_class in the automated_report factory to a report
that does not need source to be set.

Alternatively, the factory COULD also set source ... but it would have to be
explicitly set back to nil when creating an instance with a report type that
does NOT need source (to pass the :source_must_be_valid_for_report_type validation)

Easier to use a report type that does not need source as the default (FederationGrowthReport).