SessionSource: spec: add source to Compl+Fed reps spec
Add source to 'Compliance Reports' and 'Federation Reports' feature spec.

Add support for receiving source to support code in the
'Subscribing to a nil class report' shared example.

For completeness, add this support also to the
'Subscribing to an automated report with target' shared example and to the
tests using it - even though these tests only involve reports that do NEED use
source, so in all cases, the source value passed to these tests is nil.