Correctly handle time ranges that cross year-end
When a time range spanning February and crossing the end of a calendar
year is encountered, the numeric range of months
`(start.month..finish.month)` is no longer correct since `start.month`
is typically the larger value. This caused the step size to be
miscalculated.

Using ActiveSupport's date addition stuff instead of coercing to
integers ensures that all months are handled correctly.