DB: change all t.timestamp fields to use t.datetime
Fixes #210

In MySQL, TIMESTAMP is stored as unix time integer - and even when storing a
UTC timestamp value into the field, MySQL treats it as a local timestamp in the
server timezone and applies a conversion to UTC, converting back upon retrieval.

Besides this being additional overhead, it also breaks when the UTC timestamp
stored falls on a period which does not exist in the local timezone (because of
DST change) - but is otherwise a perfectly valid UTC timestamp.

The DATETIME type does not have any of these downsides and is more suitable for
storing datetime values where the application knows what timezone they're in
(UTC in case of Reporting Service) and stops MySQL from making a (possibly
wrong) guess.

Create a new migration to convert all t.timestamp fields (across 4 tables) to t.datetime.

Update generated db/schema.rb accordingly.

Note this also drops CURRENT_TIMESTAMP defaults accidentally introduced in 37884fd.

All rspec tests pass.