DB: permit hashed_principal_name to be NULL
When an IdP is not configured with idp.fticks.salt, PN is omited from the F-Ticks record.

With the DB marking hashed_principal_name as NOT NULL, processing of such a record fails and the record is discarded.

There is no reason to make hashed_principal_name required.

F-Ticks spec [1] also explicitly says:

    4.  Common F-ticks Attributes

    In general a log consumer should not assume that any one attribute is
    present.  Depending on the situation any one of these (or any other
    defined attributes) may be missing from an F-ticks message.

So remove the NOT NULL constraint.

[1](https://tools.ietf.org/html/draft-johansson-fticks-00#page-3)