Config: use correct key for session expiry
As per https://github.com/redis-store/redis-rails
the correct key name is `expiry_after`.

When using `expire_in`, it does correctly set the Redis key expiry date,
but is ignored when constructing the cookie.

When using `expire_after`, the expiry is correctly applied to both
the Redis key and the session cookie.