Ignore invalid events from DS
Invalid imports from DS logs threw off reporting ingest. These events
are not important to us so we should simply drop them and carry on.

Specific case leading to this change was an entityID in excess of 255
char, we don't support registration of entityID that long to the request
was clearly junk from the wider internet.